backups/backup_single_gui_20250329_192853/requirements.txt
# 核心依赖
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0

# 数据处理和分析
pytz>=2021.1
python-dateutil>=2.8.2
ta-lib>=0.4.0  # 注意：这个库可能需要单独安装，请参考官方文档
statsmodels>=0.13.0
scipy>=1.7.0

# 数据获取
yfinance>=0.1.70
alpha_vantage>=2.3.1
pandas-datareader>=0.10.0
requests>=2.26.0
beautifulsoup4>=4.10.0
lxml>=4.6.3

# 数据存储
sqlite3-api>=0.1.0
sqlalchemy>=1.4.23

# 可视化
tk>=0.1.0  # 注意：这是Tkinter的包装，通常Python自带
pillow>=8.3.1
ttkthemes>=3.2.0
mplfinance>=0.12.8b0  # 用于财务图表

# 工具和辅助
tqdm>=4.62.0
joblib>=1.0.1
appdirs>=1.4.4
rich>=10.9.0
pyyaml>=6.0
python-dotenv>=0.19.0

# 开发工具（可选）
pytest>=6.2.5
black>=21.8b0
flake8>=3.9.2
mypy>=0.910 